drivers: clock_control: Separated nrf lfclk shim from nrf clock shim.
Separated clock_control_nrf_lfclk shim from clock_control_nrf shim.
