+++ boards/ezurio/bl54l15u_dvk/nrf54l15_cpuapp_common.dtsi
@@ -96,6 +96,10 @@
 	status = "okay";
 };
 
+&lfclk {
+	status = "okay";
+};
+
 &spi00 {
 	status = "okay";
 	cs-gpios = <&gpio2 5 GPIO_ACTIVE_LOW>;
