+++ boards/native/nrf_bsim/nrf54l15bsim_nrf54l15_cpuapp.dts
@@ -139,3 +139,7 @@
 &xo {
 	status = "okay";
 };
+
+&lfclk {
+	status = "okay";
+};
