+++ boards/native/nrf_bsim/nrf54lm20bsim_nrf54lm20a_cpuapp.dts
@@ -146,3 +146,7 @@
 &xo {
 	status = "okay";
 };
+
+&lfclk {
+	status = "okay";
+};
