+++ boards/nordic/nrf54l15dk/nrf54l_05_10_15_cpuapp_common.dtsi
@@ -105,6 +105,10 @@
 	status = "okay";
 };
 
+&lfclk {
+	status = "okay";
+};
+
 &gpregret1 {
 	status = "okay";
