+++ boards/nordic/nrf54lm20dk/nrf54lm20a_cpuapp_common.dtsi
@@ -135,6 +135,10 @@
 	status = "okay";
 };
 
+&lfclk {
+	status = "okay";
+};
+
 &bt_hci_controller {
 	status = "okay";
 };
