+++ boards/panasonic/panb511evb/panb511evb_nrf54l15_cpuapp_common.dtsi
@@ -95,6 +95,10 @@
 	status = "okay";
 };
 
+&lfclk {
+	status = "okay";
+};
+
 &spi00 {
 	status = "okay";
 	cs-gpios = <&gpio2 5 GPIO_ACTIVE_LOW>;
