+++ boards/we/ophelia4ev/ophelia4ev_nrf54l15_cpuapp.dts
@@ -103,6 +103,10 @@
 	status = "okay";
 };
 
+&lfclk {
+	status = "okay";
+};
+
 &spi00 {
 	status = "okay";
