+++ drivers/clock_control/Kconfig.nrf
@@ -302,21 +302,21 @@ config CLOCK_CONTROL_NRF_IRON_HSFLL_LOCAL_DVFS_TIMEOUT_MS
 
 endif # CLOCK_CONTROL_NRF_IRON_HSFLL_LOCAL
 
-config CLOCK_CONTROL_NRF_LFCLK
+config CLOCK_CONTROL_NRFS_LFCLK
 	bool "NRF LFCLK driver support"
-	depends on DT_HAS_NORDIC_NRF_LFCLK_ENABLED
+	depends on DT_HAS_NORDIC_NRFS_LFCLK_ENABLED
 	select NRFS
 	select NRFS_CLOCK_SERVICE_ENABLED
 	select CLOCK_CONTROL_NRF2_COMMON
 	default y
 
-if CLOCK_CONTROL_NRF_LFCLK
+if CLOCK_CONTROL_NRFS_LFCLK
 
-config CLOCK_CONTROL_NRF_LFCLK_CLOCK_TIMEOUT_MS
+config CLOCK_CONTROL_NRFS_LFCLK_CLOCK_TIMEOUT_MS
 	int "Timeout waiting for nrfs clock service callback in milliseconds"
 	default 1000
 
-endif # CLOCK_CONTROL_NRF_LFCLK
+endif # CLOCK_CONTROL_NRFS_LFCLK
 
 config CLOCK_CONTROL_NRF_COMMON
 	bool
@@ -330,6 +330,12 @@ config CLOCK_CONTROL_NRF_HFCLK
 config CLOCK_CONTROL_NRF_XO
 	bool "NRF XO driver support"
 	depends on DT_HAS_NORDIC_NRF_CLOCK_XO_ENABLED
+    select CLOCK_CONTROL_NRF_COMMON
+	default y
+
+config CLOCK_CONTROL_NRF_LFCLK
+	bool "NRF LFCLK driver support"
+	depends on DT_HAS_NORDIC_NRF_CLOCK_LFCLK_ENABLED
     select CLOCK_CONTROL_NRF_COMMON
 	default y
