+++ drivers/clock_control/clock_control_nrf.c
@@ -539,129 +539,6 @@ static void onoff_start(struct onoff_manager *mgr,
 	}
 }
 
-/** @brief Wait for LF clock availability or stability.
- *
- * If LF clock source is SYNTH or RC then there is no distinction between
- * availability and stability. In case of XTAL source clock, system is initially
- * starting RC and then seamlessly switches to XTAL. Running RC means clock
- * availability and running target source means stability, That is because
- * significant difference in startup time (<1ms vs >200ms).
- *
- * In order to get event/interrupt when RC is ready (allowing CPU sleeping) two
- * stage startup sequence is used. Initially, LF source is set to RC and when
- * LFSTARTED event is handled it is reconfigured to the target source clock.
- * This approach is implemented in nrfx_clock driver and utilized here.
- *
- * @param mode Start mode.
- */
-static void lfclk_spinwait(enum nrf_lfclk_start_mode mode)
-{
-	static const nrf_clock_domain_t d = NRF_CLOCK_DOMAIN_LFCLK;
-	static const nrf_clock_lfclk_t target_type =
-		/* For sources XTAL, EXT_LOW_SWING, and EXT_FULL_SWING,
-		 * NRF_CLOCK_LFCLK_XTAL is returned as the type of running clock.
-		 */
-		(IS_ENABLED(CONFIG_CLOCK_CONTROL_NRF_K32SRC_XTAL) ||
-		 IS_ENABLED(CONFIG_CLOCK_CONTROL_NRF_K32SRC_EXT_LOW_SWING) ||
-		 IS_ENABLED(CONFIG_CLOCK_CONTROL_NRF_K32SRC_EXT_FULL_SWING))
-		? NRF_CLOCK_LFCLK_XTAL
-		: CLOCK_CONTROL_NRF_K32SRC;
-	nrf_clock_lfclk_t type;
-
-	if ((mode == CLOCK_CONTROL_NRF_LF_START_AVAILABLE) &&
-	    (target_type == NRF_CLOCK_LFCLK_XTAL) &&
-	    (nrf_clock_lf_srccopy_get(NRF_CLOCK) == CLOCK_CONTROL_NRF_K32SRC)) {
-		/* If target clock source is using XTAL then due to two-stage
-		 * clock startup sequence, RC might already be running.
-		 * It can be determined by checking current LFCLK source. If it
-		 * is set to the target clock source then it means that RC was
-		 * started.
-		 */
-		return;
-	}
-
-	bool isr_mode = k_is_in_isr() || k_is_pre_kernel();
-	int key = isr_mode ? irq_lock() : 0;
-
-	if (!isr_mode) {
-		nrf_clock_int_disable(NRF_CLOCK, NRF_CLOCK_INT_LF_STARTED_MASK);
-	}
-
-	while (!(nrfx_clock_is_running(d, (void *)&type)
-		 && ((type == target_type)
-		     || (mode == CLOCK_CONTROL_NRF_LF_START_AVAILABLE)))) {
-		/* Synth source start is almost instant and LFCLKSTARTED may
-		 * happen before calling idle. That would lead to deadlock.
-		 */
-		if (!IS_ENABLED(CONFIG_CLOCK_CONTROL_NRF_K32SRC_SYNTH)) {
-			if (isr_mode || !IS_ENABLED(CONFIG_MULTITHREADING)) {
-				k_cpu_atomic_idle(key);
-			} else {
-				k_msleep(1);
-			}
-		}
-
-		/* Clock interrupt is locked, LFCLKSTARTED is handled here. */
-		if ((target_type ==  NRF_CLOCK_LFCLK_XTAL)
-		    && (nrf_clock_lf_src_get(NRF_CLOCK) == NRF_CLOCK_LFCLK_RC)
-		    && nrf_clock_event_check(NRF_CLOCK,
-					     NRF_CLOCK_EVENT_LFCLKSTARTED)) {
-			nrf_clock_event_clear(NRF_CLOCK,
-					      NRF_CLOCK_EVENT_LFCLKSTARTED);
-			nrf_clock_lf_src_set(NRF_CLOCK,
-					     CLOCK_CONTROL_NRF_K32SRC);
-
-			/* Clear pending interrupt, otherwise new clock event
-			 * would not wake up from idle.
-			 */
-			NVIC_ClearPendingIRQ(DT_INST_IRQN(0));
-			nrf_clock_task_trigger(NRF_CLOCK,
-					       NRF_CLOCK_TASK_LFCLKSTART);
-		}
-	}
-
-	if (isr_mode) {
-		irq_unlock(key);
-	} else {
-		nrf_clock_int_enable(NRF_CLOCK, NRF_CLOCK_INT_LF_STARTED_MASK);
-	}
-}
-
-void z_nrf_clock_control_lf_on(enum nrf_lfclk_start_mode start_mode)
-{
-	static atomic_t on;
-	static struct onoff_client cli;
-
-	if (atomic_set(&on, 1) == 0) {
-		int err;
-		struct onoff_manager *mgr =
-				get_onoff_manager(CLOCK_DEVICE,
-						  CLOCK_CONTROL_NRF_TYPE_LFCLK);
-
-		sys_notify_init_spinwait(&cli.notify);
-		err = onoff_request(mgr, &cli);
-		__ASSERT_NO_MSG(err >= 0);
-	}
-
-	/* In case of simulated board leave immediately. */
-	if (IS_ENABLED(CONFIG_SOC_SERIES_BSIM_NRFXX)) {
-		return;
-	}
-
-	switch (start_mode) {
-	case CLOCK_CONTROL_NRF_LF_START_AVAILABLE:
-	case CLOCK_CONTROL_NRF_LF_START_STABLE:
-		lfclk_spinwait(start_mode);
-		break;
-
-	case CLOCK_CONTROL_NRF_LF_START_NOWAIT:
-		break;
-
-	default:
-		__ASSERT_NO_MSG(false);
-	}
-}
-
 static void clock_event_handler(nrfx_clock_evt_type_t event)
 {
 	const struct device *dev = CLOCK_DEVICE;
