+++ dts/arm/nordic/nrf51822.dtsi
@@ -80,6 +80,13 @@
 			status = "okay";
 		};
 
+        lfclk: lfclk@40000000 {
+            compatible = "nordic,nrf-clock-lfclk";
+            reg = <0x40000000 0x1000>;
+            interrupts = <5 NRF_DEFAULT_IRQ_PRIORITY>;
+            status = "okay";
+        };
+
 		nrf_mpu: nrf-mpu@40000000 {
 			compatible = "nordic,nrf-mpu";
 			reg = <0x40000000 0x1000>;
