+++ dts/arm/nordic/nrf52832.dtsi
@@ -72,6 +72,13 @@
 			status = "okay";
 		};
 
+        lfclk: lfclk@40000000 {
+            compatible = "nordic,nrf-clock-lfclk";
+            reg = <0x40000000 0x1000>;
+            interrupts = <0 NRF_DEFAULT_IRQ_PRIORITY>;
+            status = "okay";
+        };
+
 		power: power@40000000 {
 			compatible = "nordic,nrf-power";
 			reg = <0x40000000 0x1000>;
