+++ dts/arm/nordic/nrf5340_cpunet.dtsi
@@ -69,6 +69,13 @@
 			status = "okay";
 		};
 
+        lfclk: lfclk@41005000 {
+            compatible = "nordic,nrf-clock-lfclk";
+            reg = <0x41005000 0x1000>;
+            interrupts = <5 NRF_DEFAULT_IRQ_PRIORITY>;
+            status = "okay";
+        };
+
 		power: power@41005000 {
 			compatible = "nordic,nrf-power";
 			reg = <0x41005000 0x1000>;
