+++ dts/arm/nordic/nrf91_peripherals.dtsi
@@ -351,6 +351,13 @@ hfclk: hfclk@5000 {
 	status = "okay";
 };
 
+lfclk: lfclk@5000 {
+	compatible = "nordic,nrf-clock-lfclk";
+	reg = <0x5000 0x1000>;
+	interrupts = <5 NRF_DEFAULT_IRQ_PRIORITY>;
+	status = "okay";
+};
+
 power: power@5000 {
 	compatible = "nordic,nrf-power";
 	reg = <0x5000 0x1000>;
