+++ dts/bindings/clock/nordic,nrf-clock-lfclk.yaml
@@ -1,0 +1,15 @@
+# Copyright (c) 2025 Nordic Semiconductor ASA
+# SPDX-License-Identifier: Apache-2.0
+
+description: Nordic nRF lfclk clock control node
+
+compatible: "nordic,nrf-clock-lfclk"
+
+include: base.yaml
+
+properties:
+  reg:
+    required: true
+
+  interrupts:
+    required: true
