+++ tests/drivers/clock_control/clock_control_api/src/nrf_device_subsys.h
@@ -7,34 +7,31 @@
 #include "device_subsys.h"
 #include <zephyr/drivers/clock_control/nrf_clock_control.h>
 
-
-#ifndef CONFIG_SOC_NRF52832
-static const struct device_subsys_data subsys_data[] = {
-	/* On nrf52832 LF clock cannot be stopped because it leads
-	 * to RTC COUNTER register reset and that is unexpected by
-	 * system clock which is disrupted and may hang in the test.
-	 */
-	{
-		.subsys = CLOCK_CONTROL_NRF_SUBSYS_LF,
-		.startup_us = (CLOCK_CONTROL_NRF_K32SRC ==
-			NRF_CLOCK_LFCLK_RC) ? 1000 : 500000
-	}
-};
-#endif /* !CONFIG_SOC_NRF52832 */
-
+#if NRF_CLOCK_HAS_HFCLK   
 static const struct device_subsys_data subsys_data_hfclk[] = {
 	{
 		.subsys = CLOCK_CONTROL_NRF_SUBSYS_HF,
 		.startup_us = CONFIG_TEST_NRF_HF_STARTUP_TIME_US
 	}
 };
-
+#endif // NRF_CLOCK_HAS_HFCLK
+#if NRF_CLOCK_HAS_XO
 static const struct device_subsys_data subsys_data_xo[] = {
 	{
 		.subsys = CLOCK_CONTROL_NRF_SUBSYS_HF,
 		.startup_us =  CONFIG_TEST_NRF_HF_STARTUP_TIME_US
 	}
 };
+#endif // NRF_CLOCK_HAS_XO
+#if !defined(CONFIG_SOC_NRF52832)
+static const struct device_subsys_data subsys_data_lfclk[] = {
+	{
+		.subsys = CLOCK_CONTROL_NRF_SUBSYS_LF,
+		.startup_us = (CLOCK_CONTROL_NRF_K32SRC ==
+			    NRF_CLOCK_LFCLK_RC) ? 1000 : 500000
+	}
+};
+#endif // !defined(CONFIG_SOC_NRF52832)
 
 static const struct device_data devices[] = {
 #if NRF_CLOCK_HAS_HFCLK    
@@ -52,10 +49,14 @@ static const struct device_data devices[] = {
     },
 #endif // NRF_CLOCK_HAS_XO
 #if !defined(CONFIG_SOC_NRF52832)
+	/* On nrf52832 LF clock cannot be stopped because it leads
+	 * to RTC COUNTER register reset and that is unexpected by
+	 * system clock which is disrupted and may hang in the test.
+	 */
 	{
-		.dev = DEVICE_DT_GET_ONE(nordic_nrf_clock),
-		.subsys_data =  subsys_data,
-		.subsys_cnt = ARRAY_SIZE(subsys_data)
-	}
+		.dev = DEVICE_DT_GET_ONE(nordic_nrf_clock_lfclk),
+		.subsys_data =  subsys_data_lfclk,
+		.subsys_cnt = ARRAY_SIZE(subsys_data_lfclk)
+	},
 #endif // !defined(CONFIG_SOC_NRF52832)
 };
